Added support for custom rule link
Using the environment variable 'EFF_RULE_SEARCH_LINK' it's now possible
to use a different website to link rules to that are not part of the
standard ESLint rules. It also means you can link to a different search
engine than Google. Google is kept as fallback when the environment
variable isn't set.