Added env var support for grouping by issue
Using the environment variable 'EFF_BY_ISSUE' it's now possible to
group the lint output by issue just like the passed argument 'eff-by-
issue'. The passed argument takes precedence, meaning if the argument is
given, the environment variable is ignored. If not, the value of the
environment variable dictates the behavior.