Added env var support for filtering by issue
Using the environment variable 'EFF_FILTER' it's now possible to only
output errors/warnings for a specific rule given just like the passed
CLI option 'eff-filter'. The CLI option takes precedence, meaning if the
option is given, the environment variable is ignored. If not, the value
of the environment variable dictates the behavior.